Added environment variable check to config.py
If the required keys are all in os.environ, then
config.py does not bother with env.yaml and returns
immediately.

This change was made so that the app could work on
Heroku (or other similar environments). Since the
env.yaml file will not be deployed, the configuration
settings must be obtained from environment variables
in the prod environment.