Added is_debug method to config.py
The application environment is determined by whether
the config settings are obtained from os.environ
or the env.yaml file.

If it's the env.yaml file, the app is running in
debug mode (SqlAlchemy echo=True, Flask debug=False).

If it's os.environ, the app is running in prod mode.